requirements.txt
# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
alembic>=1.13.1

# Validation and settings
pydantic>=2.5.3
pydantic-settings>=2.1.0
email-validator>=2.1.0

# Authentication
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
python-multipart>=0.0.6

# AWS S3
boto3>=1.34.0

# Task queue
celery>=5.3.6
redis>=5.0.1
flower>=2.0.1

# Utilities
python-dotenv>=1.0.0

# Development and testing
pytest>=7.4.4
pytest-asyncio>=0.23.3
httpx>=0.26.0
